backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
psycopg2-binary
python-dotenv
alembic
pytest
